StormpathBaseUser should be abstract model
Fixes: https://github.com/stormpath/stormpath-django/issues/2

Since both StormpathBaseUser and StormpathUser use PermissionsMixin
which hardcodes the related names for the user model, if
StormpathBaseUser isn't abstract, the related names clash.

This bug was hidden in Django 1.5 because the PermissionsMixin
didn't hardcode related names, but was exposed in 1.6.x.

We should also probably not supply ready-made StormpathUser because
it will clash with any user-customized StormpathBaseUser subclass
for the same reason.