We can't run the live tests on pull requests
This disables them on pull requests so they don't even run.
Otherwise they will report failing when it's only a matter of travis
not exposing the secure env vars to PRs. We don't wan't to get used
to seeing the little red crosses as we'll probably ignore the test
when they do fail for a legitimate reason.