Added migrations for all supported Django versions.
We need south_migrations directory with south migrations for Django 1.6 (https://docs.djangoproject.com/en/1.8/topics/migrations/#libraries-third-party-apps).
Also, for Django 1.7 and Django 1.8, we have to use migrations generated on Django 1.7 version (https://docs.djangoproject.com/en/1.8/topics/migrations/#supporting-multiple-django-versions). It should work on Django 1.8 (as stated in docs), but it doesn't - the last_login field was made nullable in 1.8. I changed that field in the migration, so it will work for both versions.